pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "supabase>=2.11.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.4.0",
]

[project.scripts]
dev = "fastapi:app"

[dependency-groups]
dev = [
    "pytest-asyncio>=0.25.1",
    "pytest>=8.3.4",
    "pytest-env>=1.1.5",
    "pre-commit>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
